+++ pkgs/sdk/server-ai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.9.3](https://github.com/launchdarkly/dotnet-core/compare/LaunchDarkly.ServerSdk.Ai-v0.9.2...LaunchDarkly.ServerSdk.Ai-v0.9.3) (2026-03-05)
+
+
+### Bug Fixes
+
+* Make defaultValue optional with a disabled default ([#232](https://github.com/launchdarkly/dotnet-core/issues/232)) ([f69d420](https://github.com/launchdarkly/dotnet-core/commit/f69d42034bb960f83b831d2edf6788f70a20ceed))
+
 ## [0.9.2](https://github.com/launchdarkly/dotnet-core/compare/LaunchDarkly.ServerSdk.Ai-v0.9.1...LaunchDarkly.ServerSdk.Ai-v0.9.2) (2026-02-25)